Pulled robot animation out into another file
