+++ next.config.mjs
@@ -1,6 +1,9 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
-  output: "standalone"
+  output: "standalone",
+  images: {
+    formats: ['image/webp'],
+  },
 }
 
 export default nextConfig
